Align viewport to tile pixels
All tile coordinates are now guaranteed to be integers. Fine detail in
the tile images, such as very thin lines, are now preserved in the PNGs,
where previously they were slightly distorted by the sub-pixel
translation of the tiles.

Also no longer include any tiles that are outside of the viewport
(previously there could be an extra row at the bottom or column at the
right, if the viewport ended at a tile boundary).